Add tox and other CI tools.
There was still a problem of missing HISTORY.md and README.md when
running tests with tox. Tox is installing from .zip that I believe is *the
python distribution package*, not from local directory which solely used pip
is doing. I think this was the reason why the #56 was closed though I've hit
this.

Fixes #56