Turn host.session into a property
Its possible to end up calling execute on a host instance with no
session, resulting in an attribute error.  I also intermittently would
even get a TypeError because of an empty session when calling host.connect(),
but could not find a consistent reproducer.

Since connect is handled automatically along with the session property,
there is no longer a need to track connected state through pickling, so
`__setstate` has been removed.