Refactor Provider registration system
This new version is more in-line with a plugin-based approach.
A Provider author can now create a new Provider class and register it
without having to touch any other files but their own.

To do this, we now have automatic registration of any class that inherits
from the base Provider class. Additionally, registering actions is done via
a new decorator.