Add contextual awareness to host creation
This change fixes an issue we're seeing where host entries, missing a
hostname and ip, are being reconstructed for checkin. While that shouldn't
be an issue during checkin, it would be if used for other purposes.

With that in mind, this adds a check to see if the host is being
reconstructed. If so, then it will just debug log instead of raising an
error.