Swap Broker formatting from black to ruff
uff now offers code formatting that is largely black-compatible, with some deviations [1].
Since we are already using Ruff for code linting in the project, we can also rely on it to handle code formatting.
This will not only decrease the number of dependencies, but also be faster and (arguably) better formatted.

[1] - https://docs.astral.sh/ruff/formatter/black/