Attempt to fix broker quay image build/push
The quay publish job is having trouble installing broker in the
container. I simplified the Dockerfile a bit and things are building
correctly locally.

I also improved the interactive logic with the settings file download,
this was initially reproduced by running the container for `--version`
while not passing interactive flags to docker.