Split ssh2-python into an extra dependency (#264)
The intent of this change is to get us back to publishing Broker to
PyPi. In order to do this, I split the ssh2-python github dependency
from the main dependencies and into the the extras.
As part of this, this also allows users to choose their ssh2-python
backend (3.12 fork or 3.11 and earlier main package).