Add full compatibility for podman
This change fills the gap previously seen for podman's inability execute
commands on containers.
A large part of this was handling the returned output from podman's
exec_run, which is a tuple containing the return code and a bytes string
containing both the stdout and stderr.
Due to this, we needed to implement a demuxer.

Also made some minor tweaks to the functional tests