Add ability to specify multiple container networks AND provider instance
This change allows users to use a comma-separated list in the container
network settings. This will then expand to pass each network on
container creation. If any of the networks aren't found, then an
exception will be raised.

I also threw in the ability to specify an alternate provider instance to
load when doing most broker actions, like checkout.