Updated uv usage and documentation
With recent changes to uv, I am solidifying my position for the use of
it over pip.
Part of this includes changing the usage of uv in our actions as well as
recomending it in our documentation.

Since most Broker users are no longer also Broker developers, I've
changed the quickstart to not involve local development.