Introduction of broker config cli and much more!
The initial change of this was to introduce the broker config cli.
However, to do this, a number of supporting and related changes needed
to be made.

The first of these is the new ConfigManager class that manages reading
from and writing to the config file.
Since so much logic was implemented for this, it made more sense to roll
in the import/init functionality, that was originally in the settings
module, into the new class.

This also gave me the opportunity to separate the test config from the
example config. This removal, and other changes in this were supported
by ConfigManager's migration functionality.

To simplify the nested chunk notation, I moved provider instances out
from a list to just being a nested dictionary. This actually simplified
the instance logic quite a bit.

Related to config simlification was the separation of the host settings
to a new ssh chunk of the config. This is because not all users use the
host functionality Broker provides. A likely change with this release is
going to be removing a default ssh backend from Broker's requirements.

While I was at it, I also switched the CLI over to rich-click as part of
a larger effort to improve the CLI experience.