Rework inventory cli
With the changes we've been making in 0.6 toward a better, more
customizable and visually appealing, user experience. It's time that we
revisited the inventory command.
The --curated view introduced in 0.5.x has now become the default
inventory view. The original default view is now present in its new form
and home with the --list flag.
The --details flag largely retains the functionality it did before.
However, it involves much less processing than before.

One important piece added in this is the ability to customize the
inventory fields on a per-user basis. The details on this customization
are documented in detail in the supporting code.
It will also be important to document this in the wiki upon release.