Fix issue around host garbage collection
As identified in #334, we can get into a state during garbage collection
where a Host instance's _session attribute is missing.
However, the del method calls close, which expects that attribute to be
in place.
In this change, we add a couple of conditions to the del to gate the
close call.

fixes #334