pyproject.toml updates
Now that ssh2-python has returned, let's switch the primary ssh-backend
dependency to it. This primary dependency will be removed in Broker 0.7
along with the ssh2_python311 and ssh2_python312 optional dep
definitions.

Also, with the merge of #346, we can fix how the license is attributed
in the project metadata. This moves us in compliance with the newer PEP
639 standard.

Fixes #347