Update the way we're getting a container's hostname (#352)
Previously, if a user set a hostname on a container, Broker wouldn't
recognize that. This is because we just looked up the container's id.
When a hostname isn't set at runtime, this is perfectly fine.
However, when a user does set a hostname, this breaks what would be
expected.

With this change, we have a more reliable way to solve both cases.