Allow checkout errors to propagate correctly
This change is primarily based around propagating errors in a more
correct way. Previously, we stripped them out during host construction,
which was masked due to the errors still being logged on init.

Now, we keep the errors how they should be and raise the latest.
This also combines a the loggedcli with the ExceptionHandler class in
the cli to reduce redundancy.

Finally, we're not going to re-log the exception message as the command
exits.