Verify users have access to the default AnsibleTower inventories
When a user call a workflow (or job template) without passing in an
inventory, Broker will not pass a inventory and let ansible figure out
what to do.
However, this can lead to bad behavior when the user doesn't have access
to that default inventory, or there is not default set.

To remedy this, we will perform some basic inventory validation to make
sure a default is available and that the user can at least view the
inventory.