Switch build backend to uv_build
This change swaps our build backend from setuptools to uv_build.
A downside of this is that we lose scm-based versioning.
To accomodate for that, I've also added a helper tool `bumpver` to the
dev dependencies (thanks to Ondrej's suggestion) that makes it easy to
bump versions.

```
bumpver update --patch
or
bumpver update --major
```

Also switching to uv publish while we're a it.